requirements.txt
python-hcl2
rich
streamlit
pandas